backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
Werkzeug==3.0.3
requests>=2.31.0
ragflow-sdk>=0.1.0
openai>=1.0.0
